requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
alembic
pydantic>=2.0.0
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
numpy
scikit-learn
shapely
geopandas
folium
python-dotenv
asyncpg
psycopg2-binary
aiosqlite
httpx

